Rename call expression flattener to match canonical declaration
