+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/type-script-ast/flattenCallExpressionChain.ts
